+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-04-21
+
+### Fixed
+
+- Fixed bug where the CSV Export on the Item Tracker would only export items for the selected category rather than all items.
+
 ## 2024-04-20
 
 ### Fixed